Blob Recorder and Video Widget
Turns out that both video and audio recording builds off blob
So we get another level of abstraction - no need for separate recorders!

Had to make more changes due to memoization logic and things across the
board in order to support the idea of re-creating machines when settings
change. It's working well :)